Revert "drop python-dev  dep"
This reverts commit 39cd24496ffa32e10f588fe21c4695a27014ee3d.

Python-dev build dependency needs to be set back on kata-qemu-lite as it is not
possible to build the package.
The following error is displayed when attempting to build without the
dependency:

ERROR: Cannot use '/usr/bin/python', Python 2.6 or later is required.
       Note that Python 3 or later is not yet supported.
       Use --python=/path/to/python to specify a supported Python.

kata-qemu-lite package used in Clear Linux OS is based on the one released by
Kata Containers [1], where a python3-compatible version is not yet available;
there is no equivalent package from qemu [2].

We are waiting for the Kata Containers team to provide a python3 compatible
package (WIP), but in the meantime this dependency will be set back again.

[1] https://github.com/kata-containers/qemu/tree/qemu-lite-2.11.0
[2] https://github.com/qemu/qemu

Signed-off-by: Daniela Plascencia <daniela.plascencia@linux.intel.com>